+++ server/migrations/index.js
@@ -7,4 +7,5 @@ module.exports = [
   require('./001_platform_schema_patches'),
   require('./002_embedding_vector_dimensions'),
   require('./003_system_settings_data_patches'),
+  require('./004_organizations_description'),
 ];